mod: fix crash with et.gentity_set on 64-bit windows (#2673)
* mod: fix crash with et.gentity_set on 64-bit windows

Also patch both get/set functions from tvgame lua, which were both unpatched

* game: fix mapping field size in gentity_field_t

* tvgame: fix mapping size on tvgame too

+++ src/tvgame/tvg_lua.h
@@ -103,7 +103,7 @@ typedef struct tvgame_field_s
 {
 	const char *name;
 	int type;
-	unsigned long mapping;
+	uintptr_t mapping;
 	int flags;
 } tvgame_field_t;
